fixing error related to download and refactoring
